+++ src/docfx.website.themes/default/ManagedReference.extension.js
