+++ src/docfx.website.themes/default/ManagedReference.html.primary.js
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 var mrefCommon = require('./ManagedReference.common.js');
-var extension = require('./ManagedReference.html.primary.extension.js')
+var extension = require('./ManagedReference.extension.js')
 
 exports.transform = function (model) {
   model = extension.preSteps(model);
