+++ src/docfx.website.themes/default/RestApi.html.primary.js
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 var restApiCommon = require('./RestApi.common.js');
-var extension = require('./RestApi.html.primary.extension.js')
+var extension = require('./RestApi.extension.js')
 
 exports.transform = function (model) {
   model = extension.preSteps(model);
